[Update] BTTV Emote Extensions (#1921)
* Delete extensions/bttv-emote directory

* Add files via upload

* Create .eslintrc.json

* Create .prettierrc

* Update .prettierrc

* Update .prettierrc

* Update .prettierrc

* Update .prettierrc

* Update CHANGELOG.md
