+++ extensions/bttv-emote/CHANGELOG.md
@@ -1,0 +1,9 @@
+# BTTV Emote Changelog
+
+ ## [Introducing GRID View] - 2022-06-09
+
+ - Added new grid layout for emotes.
+ - UI / UX improvement.
+ - Added empty emote view.
+ - Updated packages to latest versions.
+ - Updated metadata.
